+++ src/main/java/mdteam/ait/core/blocks/MonitorBlock.java
@@ -7,10 +7,7 @@ import mdteam.ait.core.AITDimensions;
 import mdteam.ait.core.blockentities.ConsoleBlockEntity;
 import mdteam.ait.core.blockentities.MonitorBlockEntity;
 import mdteam.ait.core.blocks.types.HorizontalDirectionalBlock;
-import net.minecraft.block.Block;
-import net.minecraft.block.BlockEntityProvider;
-import net.minecraft.block.BlockState;
-import net.minecraft.block.ShapeContext;
+import net.minecraft.block.*;
 import net.minecraft.block.entity.BlockEntity;
 import net.minecraft.block.entity.BlockEntityTicker;
 import net.minecraft.block.entity.BlockEntityType;
@@ -18,14 +15,21 @@ import net.minecraft.entity.Entity;
 import net.minecraft.entity.ItemEntity;
 import net.minecraft.entity.LivingEntity;
 import net.minecraft.entity.player.PlayerEntity;
+import net.minecraft.item.ItemPlacementContext;
 import net.minecraft.item.ItemStack;
 import net.minecraft.particle.ParticleTypes;
+import net.minecraft.state.StateManager;
+import net.minecraft.state.property.IntProperty;
+import net.minecraft.state.property.Properties;
 import net.minecraft.util.ActionResult;
+import net.minecraft.util.BlockMirror;
+import net.minecraft.util.BlockRotation;
 import net.minecraft.util.Hand;
 import net.minecraft.util.function.BooleanBiFunction;
 import net.minecraft.util.hit.BlockHitResult;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.Direction;
+import net.minecraft.util.math.RotationPropertyHelper;
 import net.minecraft.util.shape.VoxelShape;
 import net.minecraft.util.shape.VoxelShapes;
 import net.minecraft.world.BlockRenderView;
@@ -39,10 +43,15 @@ import java.util.Random;
 
 import static mdteam.ait.core.blocks.DoorBlock.rotateShape;
 
-public class MonitorBlock extends HorizontalDirectionalBlock implements BlockEntityProvider {
+public class MonitorBlock extends BlockWithEntity implements BlockEntityProvider {
+    public static final int MAX_ROTATION_INDEX = RotationPropertyHelper.getMax();
+    private static final int MAX_ROTATIONS = MAX_ROTATION_INDEX + 1;
+    public static final IntProperty ROTATION = Properties.ROTATION;
+    protected static final VoxelShape SHAPE = Block.createCuboidShape(4.0, 0.0, 4.0, 12.0, 8.0, 12.0);
 
     public MonitorBlock(Settings settings) {
         super(settings);
+        this.setDefaultState(this.stateManager.getDefaultState().with(ROTATION, 0));
     }
 
     public VoxelShape eastShape() {
@@ -58,12 +67,12 @@ public class MonitorBlock extends HorizontalDirectionalBlock implements BlockEnt
 
     @Override
     public VoxelShape getCollisionShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
-        return rotateShape(Direction.EAST, state.get(FACING), eastShape());
+        return SHAPE;
     }
 
     @Override
     public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
-        return rotateShape(Direction.EAST, state.get(FACING), eastShape());
+        return SHAPE;
     }
 
 
@@ -90,5 +99,30 @@ public class MonitorBlock extends HorizontalDirectionalBlock implements BlockEnt
             monitorBlockEntity.useOn(world, player.isSneaking(), player);
 
         return ActionResult.SUCCESS;
+    }
+
+    @Override
+    public VoxelShape getCullingShape(BlockState state, BlockView world, BlockPos pos) {
+        return VoxelShapes.empty();
+    }
+
+    @Override
+    public BlockState getPlacementState(ItemPlacementContext ctx) {
+        return this.getDefaultState().with(ROTATION, RotationPropertyHelper.fromYaw(ctx.getPlayerYaw()));
+    }
+
+    @Override
+    public BlockState rotate(BlockState state, BlockRotation rotation) {
+        return state.with(ROTATION, rotation.rotate(state.get(ROTATION), MAX_ROTATIONS));
+    }
+
+    @Override
+    public BlockState mirror(BlockState state, BlockMirror mirror) {
+        return state.with(ROTATION, mirror.mirror(state.get(ROTATION), MAX_ROTATIONS));
+    }
+
+    @Override
+    protected void appendProperties(StateManager.Builder<Block, BlockState> builder) {
+        builder.add(ROTATION);
     }
 }