+++ src/main/java/mdteam/ait/core/commands/LinkCommand.java
@@ -1,0 +1,43 @@
+package mdteam.ait.core.commands;
+
+import com.mojang.brigadier.Command;
+import com.mojang.brigadier.CommandDispatcher;
+import com.mojang.brigadier.arguments.StringArgumentType;
+import com.mojang.brigadier.context.CommandContext;
+import mdteam.ait.AITMod;
+import mdteam.ait.api.tardis.LinkableItem;
+import mdteam.ait.tardis.Tardis;
+import mdteam.ait.tardis.wrapper.server.manager.ServerTardisManager;
+import net.minecraft.command.argument.UuidArgumentType;
+import net.minecraft.item.ItemStack;
+import net.minecraft.server.command.ServerCommandSource;
+import net.minecraft.server.network.ServerPlayerEntity;
+import net.minecraft.sound.SoundEvent;
+
+import static mdteam.ait.core.commands.TeleportInteriorCommand.TARDIS_SUGGESTION;
+import static net.minecraft.server.command.CommandManager.argument;
+import static net.minecraft.server.command.CommandManager.literal;
+
+public class LinkCommand {
+    public static void register(CommandDispatcher<ServerCommandSource> dispatcher) {
+        dispatcher.register(literal(AITMod.MOD_ID)
+                .then(literal("link").requires(source -> source.hasPermissionLevel(2))
+                        .then(argument("tardis", UuidArgumentType.uuid()).suggests(TARDIS_SUGGESTION)
+                                    .executes(LinkCommand::runCommand))));
+    }
+
+    private static int runCommand(CommandContext<ServerCommandSource> context) {
+        ServerPlayerEntity source = context.getSource().getPlayer();
+        Tardis tardis = ServerTardisManager.getInstance().getTardis(UuidArgumentType.getUuid(context, "tardis"));
+
+        if (tardis == null || source == null) return 0;
+
+        ItemStack stack = source.getMainHandStack();
+
+        if (!(stack.getItem() instanceof LinkableItem linker)) return 0;
+
+        linker.link(stack, tardis);
+
+        return Command.SINGLE_SUCCESS;
+    }
+}