+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.2.26
 - Autopilot now only works on lowest speed
+- Flipping the handbrake with autopilot enabled now forces you to land instead of crashing
 - Updated monitor screen
 - Added eighth doctor hum
 - Made a different particle show when refueling in a RiftChunk