+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Failing to takeoff/land no longer sends a message to the pilot
 - Flight sounds no longer play when out of power
 - Fixed being unable to turn off antigravs at max world height
+- Fixed issues with Siege Mode
 - Created ArtronHolder interface
 - Renamed "ExteriorSchema" to "ExteriorCategory"
 - Moved "hasPortals" method into the individual variants