+++ CHANGELOG.md
@@ -1,3 +1,2 @@
-# 1.2.28
-- People who do not have a key when security is enabled cannot open the monitor from the monitor block anymore
-- Hot fix crash
+# 1.2.29
+- Made cloak combine alpha when taking off / landing